refactor 1. 메소드명 수정  - LadderPoints -> LadderLine  - LadderLaine -> LadderLines 2. object class사용  - Objects.isNull
