+++ src/main/java/ladder/domain/LadderGame.java
@@ -2,23 +2,23 @@ package ladder.domain;
 
 
 public class LadderGame {
-    private Persons person;
-    private LadderLine ladderLine;
+    private Persons persons;
+    private LadderLines ladderLine;
 
     public LadderGame(final String names, final int ladderHeight) {
         initLadderGame(names, ladderHeight);
     }
 
     private void initLadderGame(final String names, final int ladderHeight) {
-        this.person = new Persons(names);
-        this.ladderLine = new LadderLine(ladderHeight, person.getPersons().size());
+        this.persons = new Persons(names);
+        this.ladderLine = new LadderLines(ladderHeight, persons.getPersons().size());
     }
 
     public Persons getPerson() {
-        return person;
+        return persons;
     }
 
-    public LadderLine getLadderLine() {
+    public LadderLines getLadderLine() {
         return ladderLine;
     }
 }
