+++ src/test/java/domain/LadderLineTest.java
@@ -1,6 +1,6 @@
 package domain;
 
-import ladder.domain.LadderLine;
+import ladder.domain.LadderLines;
 import ladder.util.ErrorMessage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -10,14 +10,14 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 public class LadderLineTest {
 
-    private static LadderLine LadderLineTest;
+    private static LadderLines LadderLinesTest;
 
     @ParameterizedTest
     @DisplayName("입력받은 사다리 높이가 0 또는 음수일 때 예외를 던진다")
     @ValueSource(ints = {0, -100, -200, -3})
     void checkLadderHeight(int ladderHeight){
         assertThatThrownBy(() ->
-                LadderLineTest = new LadderLine(ladderHeight, 10)
+                LadderLinesTest = new LadderLines(ladderHeight, 10)
         ).isInstanceOf(RuntimeException.class)
                 .withFailMessage(ErrorMessage.getCheckInputLadderHeight());
     }
